Added patch from bug 2307812:
My patch for bug 2284794 introduced another bug. Oracle doesn't like
precision on the CLOB data type. It fails with "missing right parenthesis"
error. I thought that it would make sense to introduce a DataType class,
that can return more than just a text string for a DataType, in this
instance "supportsPrecision". It requires modifying the signatures of all
the getXXXType methods in the Database interface, and all its
implementations, and changing the body of the methods to return a DataType
instance. It's a fairly large change, but I believe it will allow greater
extensibility

git-svn-id: http://liquibase.jira.com/svn/CORE/trunk@858 e6edf6fb-f266-4316-afb4-e53d95876a76