Allow namespaced attributes on the following elements:  - <databaseChangeLog> - semantically, applying to this whole changelog.  - <include> - semantically, applying to the whole included changelog.  - <includeAll> - as with <include>  - <changeSet> - semantically, applying only to that change set.
Extension attributes must be from a namespace other than
http://www.liquibase.org/xml/ns/dbchangelog, and should be defined by an XML
schema.

This is largely intended to allow external tools to grab onto information
contained in (XML) changelog files. As such, none of these attributes are
exposed inwards to the rest of liquibase (and they're not visible to
extensions... though it'd be possible to expose them).