Added fix for reading column size in H2.
On my Ubuntu box, H2 sets the column name for the size of a column to CHARACTER_MAXIMUM_LENGTH and the label to COLUMN_SIZE.
When  mapping columns I think in general we should use the column label and not the column name.
The column label will fall back to the name if no label is specified. If one is specified then typically it's specified for a reason (standards compatibility).

Removed old h2 jar from database test drivers.