Fix for escaping index name on index creation.
Index name should be escaped on creation of an index.
Added integration test changeset for testing creation of an index using
the quoting strategy QUOTE_ALL_OBJECTS.

Simplified the escaping of objects for QUOTE_ALL_OBJECTS to quote all objects.
The original reason for smart quoting of objects is no longer needed.
It was in place when I was originally implementing this feature as I was quoting
in the correctObjectName method.