Fixes for failing unit tests in liquibase-core
This commit is rather large.
A lot of the fixes are for comparing the name for a change.
Another chunk relate to the escaping of the table name for Sql Server.

In MsSqlDatabase, there is explicit code to ignore the catalog name for escaping the table name.
The tests however expect the catalog name in the sql statement.
I'm not familiar enough with Sql Server to know what is correct.
For now, I've updated the tests to match the current output.

There was also a bug with SometimesExceptionThrowingChange not being removed from the registry.
This would cause subsequent tests to break.