(Derby) If error on determine Driver Version set a bigger
When I use liquibase with maven it can't get the Driver Version, so it is set to -1 and I can't use sequences or other resource available on bigger versions of database. I think would better set it to bigger version and not restrict the features. If you could find why it doesn't work on maven would be a solution too.