Refactored to account for variance when called on DB2i.
Per the documentation provided by IBM for DB2 z/OS and LUW
http://www.ibm.com/support/knowledgecenter/SSEPEK_11.0.0/com.ibm.db2z11.doc.java/src/tpc/imjcc_c0053013.html
https://www.ibm.com/support/knowledgecenter/SSEPGG_11.1.0/com.ibm.db2.luw.apdv.java.doc/src/tpc/imjcc_c0053013.html

The most recent refactor of the getDataServerType method returns does not account for the value returned on DB2i.
The method getDatabaseProductVersion returns a String like 07.01.0000 V7R1m0, instead of adhering to the scheme used by DB z/OS and LUW.  The method getDatabaseProductName returns DB2 UDB for AS/400 so is used to along with getDatabaseProductVersion to account for this inconsistency across DB2 platforms.