Configure change before getConfirmationMessage
When running dbDoc a custom change properties are set when
CustomChangeWrapper.validate is called. But since PR #88 already ran
changes are no longer validated for performance reasons.
But that causes problems if the properties are used in the confirmation
message (from NPEs - causing dbDoc to fail - to wrong messages - nulls
everywhere).