Print a banner line when starting from command line of via Maven plugin.
The motivation is to make is easier to identify the product version in bug reports / log files. In the future, I intend to print important environment variables and Maven properties that have a tendency to cause problems in various databases (e.g. locale settings in combination with SQL code dealing with DATE/TIMESTAMP columns etc.)