Prevent liquibase.exception.DatabaseException: Error getting jdbc:as400://xx.xx.xx/schema view with liquibase.statement.core.GetViewDefinitionStatement
At least on our system AS400JDBCDatabaseMetaData.getTables called to get views and view = null return a lot of of entries that are not really views.
For reference we use jtopen 9.3 against a v7r1 IBM i system.
The new code has the added bonus of returning OBJECT_BODY in the same query.