feat: Maven Liquibase Plugin generateChangeLog mysql sql comment supported
In a Liquibase SQL change set, you must start with the comment "--liquibase formatted sql". Unfortunately, this is not a well-formed comment in MySQL. The comment must be "-- liquibase formatted sql".

In MySQL, the --  (double-dash) comment style requires the second dash to be followed by at least one whitespace or control character (such as a space, tab, newline, and so on). This syntax differs slightly from standard SQL comment syntax, as discussed in https://dev.mysql.com/doc/refman/8.0/en/ansi-diff-comments.html Section 1.8.2.4, “'--' as the Start of a Comment”.