Tag should remain after rollback [CORE-2946]
Fixes https://liquibase.jira.com/browse/CORE-2946, making behavior of liquibase more consistent and predictable - rollback to tag behavior does not depend on whether the tag is isolated in it's own changeset or not. This also mirrors the behavior of updateToTag, so that updating to a tag, and rolling back to that tag, both have identical results.