Fixes CORE-3256 Verbose flag not getting interperated during 'status' command.
The verbose option was not recognised as such in the liquibase.integration.commandline.Main class. When recognised all options are expected to have a value but boolean options/flags do not necessarily need this. An option like --verbose does not need explicit --verbose=true, even the commandline help instructions say --verbose is enough. Fixed the value issue here as well so --verbose now works as expected.