CORE-2375: Prolonging and cleanup; use db time
Seems to be working so far, both the locking as well as the cleanup.

Still beta, hence needs to be activated with jvm command line param:

    -Dliquibase.changeLogLockAutoProlong=true

(especially in case you are using Spring Boot, you cannot configure
this in application.properties at the moment)

No matter if activated or not, will already upgrade the change log
lock table (add new columns with id of who locked, and when the
lock expires).

Changes are backward-compatible with standard lock service. However,
do better not mix them. That means, if you plan a database update,
first make sure all sercices are using this version of liquibase,
before deploying the changes.

There are new options that can be configured when using the new
lock service (also by using "-Dliquibase.<paramname> = <value>"):

- staleChangeLogLockRemovalTimeInSeconds

After how many seconds should a stale lock get removed?
In order to avoid confusion when several nodes have set
different values here, it is the service that holds the lock
that writes the time a lock expires in the database, using
the time from the database server.

Then other services just need to compare with current_timestamp
on the database, which means the stale lock removal should
always work reliably even if different clients have different
configurations, or are located / have configured different
time zones.

- changeLogLockProlongingRateInSeconds

How often should a lock get prolonged, when the prolonging
lock service is used?

Make sure that staleChangeLogLockRemovalTimeInSeconds is
generously larger than this, ideally three times or so.