Use postgres advisory lock as changelog lock service
This commit introduces a new lock service that specifically
targets Postgres >= 9.1 and uses transaction advisory lock
(pg_try_advisory_xact_lock). This fixes (at least for pg)
the issue of DATABASECHANGELOGLOCK getting stuck in lock
when Liquibase does not exit cleanly.

Transaction advisory locks have the nice property that don't need
to be explicitly released. In fact, they are automatically released
by pg when the transaction is completed or the connection closed.