Updated to properly handle symbolic links and paths with spaces.
Because spaces are converted to '%20' in uri's, paths that should have matched as relative paths were failing, and instead being loaded as absolute paths.
This change also makes the default path pull from a canonical path, which makes sure that things are resolved correctly when running from a directory with symlinks.