[CORE-3520] Add support for ALWAYS generation type with PostgreSQL auto-increment column.
When using Liquibase to create a table with an auto-incrementing integer, the corresponding identity column in PostgreSQL is always created with the constraint `BY DEFAULT`, irrespective of the value set for the generation type.

Add support to Liquibase for the `ALWAYS` generation type when creating an auto-incrementing column in PostgreSQL. This will allow the generation type to be set to `BY DEFAULT` or `ALWAYS`. The default value if the `generationType` attribute is not specified in the column tag, remains `BY DEFAULT`.