Change Liquibase#close to throw DatabaseException
There is no need to force callers to handle Exception.
The implementation of the close method can currently throw only a DatabaseException. But since the library has a parent exception class - LiquibaseException - it is used instead to not force callers to change their catch clauses if additional logic is added in the future.