Remove deprecated code in Ant tasks.
Motivation
----------

In 2014 work was done to refactor the Ant tasks and modernize them for newer
versions of Ant (refactoring changes found in commit
7679f9ea883c83129833d3e00b819c9fddd3ddc2). In the pull request for this work
it was mentioned that the deprecated code should be removed at some point.
This commit completes this work.

Modifications
-------------

The changes are largely focused on the BaseLiquibaseTask class. All methods
marked with a @Deprecated annotation were removed as well as several support
methods commented with deprecation warnings. Many of these support methods
were moved to child classes and were only there to document the change. All
supporting unit tests that were verifying legacy task support have been
removed as legacy task support is now removed. Ant version support remains
unchanged at version 1.7

Result
------

The Ant tasks continue to have the full functionality that was introduced in
2014 and the legacy task support is now removed. There is less code to support
and the BaseLiquibaseTask is simpler to read and understand.

Remove deprecated code in Ant tasks.

Motivation
----------

In 2014 work was done to refactor the Ant tasks and modernize them for newer
versions of Ant (refactoring changes found in commit
7679f9ea883c83129833d3e00b819c9fddd3ddc2). In the pull request for this work
it was mentioned that the deprecated code should be removed at some point.
This commit completes this work.

Modifications
-------------

The changes are entirely focused on the BaseLiquibaseTask class. All methods
marked with a @Deprecated annotation were removed as well as several support
methods commented with deprecation warnings. Many of these support methods
were moved to child classes and were only there to document the change. All
supporting unit tests that were verifying legacy task support have been
removed as legacy task support is now removed. Ant version support remains
unchanged at version 1.7

Result
------

The Ant tasks continue to have the full functionality that was introduced in
2014 and the legacy task support is now removed. There is less code to support
and the BaseLiquibaseTask is simpler to read and understand.

Remove deprecated code in Ant tasks.

Motivation
----------

In 2014 work was done to refactor the Ant tasks and modernize them for newer
versions of Ant (refactoring changes found in commit
7679f9ea883c83129833d3e00b819c9fddd3ddc2). In the pull request for this work
it was mentioned that the deprecated code should be removed at some point.
This commit completes this work.

Modifications
-------------

The changes are entirely focused on the BaseLiquibaseTask class. All methods
marked with a @Deprecated annotation were removed as well as several support
methods commented with deprecation warnings. Many of these support methods
were moved to child classes and were only there to document the change. All
supporting unit tests that were verifying legacy task support have been
removed as legacy task support is now removed. Ant version support remains
unchanged at version 1.7

Result
------

The Ant tasks continue to have the full functionality that was introduced in
2014 and the legacy task support is now removed. There is less code to support
and the BaseLiquibaseTask is simpler to read and understand.