DAT-8203: allow supplying a current value when prompting using UIService (#2103)
Allow specifying a currentValue when using the prompt method in the UIService. If a non-null currentValue is provided, it will be shown instead of the defaultValue. If the input handler allows empty inputs, and the user strikes enter at the prompt when a currentValue is shown, that value will be returned.