Add table name to OracleDB query
The query that is generated when retrieving all keys does not get a table name, while the code that creates the map for  the cache gets this name from the query result, which is always null. When reading the cache, the table name is used which  results in the cache always returning an empty result.