Move interactive prompting logic into core module (DAT-8642) (#2255)
This PR (in combination with the Pro PR) does a few small things:

Move the interactive CLI prompting logic into the core modules so they can be used outside of pro
Add an interface for the enum which dictates the prompting logic. This is so that we can have separate promptable enums in the core and pro code and use the same abstract prompting logic for both.