Added logic to first check if there is a liquibase_libs directory, then add to Classpath.
Added logic to first check/validate if there is a "liquibase_libs" directory AND is there are any jar files in that folder, before attempting to add "path/to/liquibase_libs/jars" to the Classpath.  This fix should work for Linux/Unix systems as well as Linux Emulators like Cygwin/GitBash for Windows.