stop database containers when spock execution ends (#2843)
The existing integration test framework was configured to stop TestSystem's when the cleanup: method was called. Spock internally called this method when it finished executing all of the tests in a class. Thus, when a particular test class was completed, its TestSystem was stopped, which in some cases was premature, since other classes might use the same TestSystem.

Instead, the TestSystem's should be stopped when Spock finishes executing all of the tests in all classes.