prefer license specified in LIQUIBASE_LICENSE_KEY edge case (DAT-10409) (#2858)
A user could specify their license key using the LIQUIBASE_PRO_LICENSE_KEY property in any provider (perhaps environment variable) and specify a second license key using the LIQUIBASE_LICENSE_KEY property in the defaults file. The defaults file value provider is not registered until after invoking the license service's license installation code, and the license service caches installed licenses so that it does not have to recheck for license keys. However, this means that in the above example, a license specified in the defaults file is not loaded because the license in the environment variable is erroneously installed and cached first.