Avoid ClassCastException when loading LogService from Scope.
Motivation
----------

When alternative Loggers are on the classpath calling for the LogService can cause a ClassCastException in the CLI. In order to avoid the exception, leveraging the interface would be preferred.

Modifications
-------------

I've altered the call for the LogService to remove the explicit cast to JavaLogService. I've added an instanceof check for the original service in order to set the parent if it needs to be.

Result
------

The result is a safe call to the LogService regardless of which logger is loaded in the CLI.