safely set code point limit for SnakeYaml (DAT-13189) (#3807)
Previously, Liquibase increased the code point limit for SnakeYaml so that particularly large snapshots could be parsed. This made use of a recently added method in the SnakeYaml codebase, so users of Liquibase who are using an old version of SnakeYaml with a newer version of Liquibase would have their execution interrupted by an exception about a missing method. This fix should catch that exception so that execution is not interrupted.