add additional information to toString/describe methods of SQLFileChange (DAT-13789) (#4059)
Quality checks uses a cache so that SQL files do not have to be read multiple times, subsequent reads will hit the cache. The key for the cache was the standard hashCode which, for SQLFileChange, did not provide sufficient entropy to avoid incorrect cache hits, because it did not include the path of the SQL file being included.