feat(build): add JReleaser support to build and package Liquibase CLI tool
This commit adds support for JReleaser to build and package Liquibase CLI tool. A new JReleaser configuration file `jreleaser.yml` was added to the root directory of the project. A new plugin `jreleaser-maven-plugin` was added to the `pom.xml` file to run JReleaser. A new step was added to the GitHub Actions workflow to run JReleaser after the build step.