Fixes issue #4413: Undocumented IT requirement: Username "liquibase"
Despite the title and description of issue #4413, the actual cause of
the failing IT was *not* an undocumented IT requirement, but the fact
that the author of the IT (unintentionally) used the static prefix
"liquibase" instead of the dynamic prefix provided by the test driver.

The solution hence is to replace the static prefix by the dynamic
prefix. Hence, there is nothing wrong with the test documentation.