feat(release-published.yml): add job to upload deb package to S3 bucket
This commit adds a new job to the release-published.yml workflow that uploads a deb package to an S3 bucket. The job runs on an Ubuntu 22.04 runner and uses Ruby 3.1.4. The deb-s3 gem is installed to enable the upload of the deb package. The package is signed with the GPG key id 60EF2CD3C14E006F.