chore: update release workflows to re-version deb package and upload it to AWS S3 bucket
The workflows have been updated to re-version the deb package and upload it to an AWS S3 bucket. The changes include adding a step to re-version the deb package, configuring AWS credentials, importing the GPG key, and uploading the deb package to the S3 bucket. These changes were made to ensure that the deb package is properly versioned and uploaded to the correct location.