chore(build.yml): remove unnecessary code for creating commit status
The code for creating a commit status in the build workflow is no longer needed and has been removed. This code was responsible for creating a commit status based on the success or failure of the Liquibase Pro build. Since this functionality is no longer required, the code has been removed to simplify the workflow.