chore(pom.xml): add 'no-test' profile to skip both integration and unit tests
The 'no-test' profile has been added to the pom.xml file. This profile sets the properties 'skip.integration.tests' and 'skip.unit.tests' to true, effectively skipping both integration and unit tests when this profile is activated. This profile can be used when running the build process and tests are not required.