chore(build-branch.yml): update workflow to only trigger on labeled p… (#5037)
The workflow file `build-branch.yml` has been updated to include the `pull_request_target` event and the `labeled` action. This means that the workflow will now be triggered when a pull request is opened, reopened, synchronized, or labeled. This change allows for more flexibility in triggering the workflow based on different events and actions.

---------

Co-authored-by: filipe <flautert@liquibase.org>