chore(release-published.yml): add SFTP upload steps for Liquibase xsds and flow-file schema
The Liquibase xsds and flow-file schema files need to be uploaded to an SFTP server for distribution. This commit adds the necessary steps to the release-published.yml workflow file to perform the SFTP uploads.

The Liquibase xsds are uploaded in two separate steps. The first step uploads the xsds from the "liquibase-core-repo/liquibase-standard/src/main/resources/www.liquibase.org/xml/ns/dbchangelog/" directory to the "/xml/ns/dbchangelog/" directory on the SFTP server. The second step uploads the xsds from the "liquibase-pro-repo/pro/src/main/resources/www.liquibase.org/xml/ns/pro/" directory to the "/xml/ns/pro/" directory on the SFTP server.

The flow-file schema is uploaded in a separate step. It uploads the schema files from the "liquibase-pro/pro/src/main/resources/schemas/" directory to the "/schemas/" directory on the SFTP server.

These SFTP upload steps are necessary to ensure that the Liquibase xsds and flow-file schema files are available for use and distribution.