chore(release-published.yml): remove unnecessary Liquibase xsds SFTP upload step
The Liquibase xsds SFTP upload step was removed as it is no longer needed for the release process. This step was responsible for uploading xsd files to the remote server using SFTP. However, it was found that these files are no longer required for the release, so the step was removed to simplify the workflow.