chore: update liquibase/build-logic workflows to v0.5.5
The liquibase/build-logic workflows in the following files have been updated to version v0.5.5:
- .github/workflows/attach-artifact-release.yml
- .github/workflows/build-nightly.yml
- .github/workflows/create-release.yml
- .github/workflows/release-published.yml
- .github/workflows/test.yml

The update includes using the latest version of the workflows to ensure compatibility and take advantage of any bug fixes or improvements.

Additionally, the pom.xml file has been updated to include SCM information for the project. This includes the connection, developer connection, URL, and tag information for the Liquibase BigQuery repository on GitHub.