DAT-9019 Automate uploading liquibase. liquibase pro xsd files to liquibase.org wordpress site (#5115)
* chore(release-published.yml): add SFTP upload steps for Liquibase xsds and flow-file schema

The Liquibase xsds and flow-file schema files need to be uploaded to an SFTP server for distribution. This commit adds the necessary steps to the release-published.yml workflow file to perform the SFTP uploads.

The Liquibase xsds are uploaded in two separate steps. The first step uploads the xsds from the "liquibase-core-repo/liquibase-standard/src/main/resources/www.liquibase.org/xml/ns/dbchangelog/" directory to the "/xml/ns/dbchangelog/" directory on the SFTP server. The second step uploads the xsds from the "liquibase-pro-repo/pro/src/main/resources/www.liquibase.org/xml/ns/pro/" directory to the "/xml/ns/pro/" directory on the SFTP server.

The flow-file schema is uploaded in a separate step. It uploads the schema files from the "liquibase-pro/pro/src/main/resources/schemas/" directory to the "/schemas/" directory on the SFTP server.

These SFTP upload steps are necessary to ensure that the Liquibase xsds and flow-file schema files are available for use and distribution.

* chore(release-published.yml): remove unnecessary Liquibase xsds SFTP upload step

The Liquibase xsds SFTP upload step was removed as it is no longer needed for the release process. This step was responsible for uploading xsd files to the remote server using SFTP. However, it was found that these files are no longer required for the release, so the step was removed to simplify the workflow.

* chore(release-published.yml): add support for configurable SFTP port for uploading files to WP Engine

* fix(release-published.yml): fix localDir path for liquibase-pro repository to point to the correct directory

The localDir path was pointing to 'liquibase-pro/pro/src/main/resources/schemas/' which was incorrect. It has been updated to 'liquibase-pro-repo/pro/src/main/resources/schemas/' to point to the correct directory.

* fix(release-published.yml): update remoteDir value from '/schemas/' to '/json/schema/' to reflect the correct directory path

* chore(release-published.yml): add support for syncing schemas directory to S3 bucket

The changes in this commit add a new command to the release-published workflow. The command syncs the `liquibase-pro-repo/pro/src/main/resources/schemas/` directory to the `s3://liquibaseorg-origin//son/schema/` S3 bucket. This allows for the schemas to be easily accessible and available for use.

* fix(release-published.yml): fix typo in S3 bucket path for syncing schemas folder

* chore(release-published.yml): update package-deb action to version v0.5.5 to include latest changes and improvements

---------

Signed-off-by: Alejandro Alvarez <jandroav@icloud.com>