Refactor XMLChangeLogSAXParser test to support duplicated changesets
This commit refactors and extends the testAllProvidedChangesetsAreLoaded test in XMLChangeLogSAXParserTest.groovy. The change aligns the test with a new feature, supporting the duplication of changeset identifiers, thereby improving the flexibility and scalability of the system. Additional conditional statements were included in the test function to distinguish between duplicated and non-duplicated changeset identifiers. Two scenarios were added in where clause to test both situations.