Replace Derby database with H2
The Apache Derby database has been replaced with H2 in the LiquibaseServletListener test class. This was done primarily because now derby is in a different package (derby-tools) that is not included with liquibase. Replaced all instances of Derby database configuration with the equivalent for H2 and renamed the related test method appropriately.