Remove Apache Derby dependency from pom.xml
The Apache Derby dependency, which was scoped for testing, has been removed from the project's POM file and the tests that used it are now using H2.
Derby package for java 8 (version 10.14.x) has a security issue (CVE-2022-46337) that is going to be fixed on series 10.17.x only that requires java 21.
Thus this testing dependency is being removed.