Global flag to restore behaviour of not failing on duplicated changeset identifiers  (#5296)
* Add global flag allowDuplicatedChangesetIdentifiers to enable Liquibase to behavior as it did before version 4.19.1 (PR #3790)

Added logic in the class to handle duplicate changeset identifiers based on a global configuration flag. This change provides an allowance for duplicated changesets but only when the flag is enabled in GlobalConfiguration. This was done to maintain backward compatibility with versions prior to 4.19.1 where ignoring duplicates was the default behavior.


* Refactor XMLChangeLogSAXParser test to support duplicated changesets

This commit refactors and extends the testAllProvidedChangesetsAreLoaded test in XMLChangeLogSAXParserTest.groovy. The change aligns the test with a new feature, supporting the duplication of changeset identifiers, thereby improving the flexibility and scalability of the system. Additional conditional statements were included in the test function to distinguish between duplicated and non-duplicated changeset identifiers. Two scenarios were added in where clause to test both situations.

* Javadocs

---------

Co-authored-by: rberezen <ruslan.berezenskyi@gmail.com>