add default constructor + setters to `RanChangeSet` & related classes
this is needed to be able to de-serialise them using [jackson] which
relies on the default constructor & setters.
in the upcoming liquibase-opensearch implementation [opensearch-java] is
used, which underneath is using jackson for the (de-)serialisation of
classes.

specifically, this is used in the implementation of
`AbstractChangeLogHistoryService#queryRanChangeSets` which needs to load
`RanChangeSet`s from the DB.

also do some minor cleanups by removing manually implemented getters &
setters where lombok annotations can do the job instead (no functional
impact).

[jackson]: https://github.com/FasterXML/jackson
[opensearch-java]: https://github.com/opensearch-project/opensearch-java/

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>