DAT-16471 DevOps: Attach artifatcs job fail (#5370)
* chore(create-release.yml): add scripts to download and sign artifacts

The commit adds two scripts, `re-version.sh` and `sign-artifacts.sh`, to the `.github/util/` directory. These scripts are downloaded from the `liquibase/liquibase` repository and made executable. The purpose of these scripts is to download and sign the artifacts for the release.

* ls

* changes

* use needs as outputs in between jobs

* fix(create-release.yml): fix indentation issue in the script execution line

* chore(create-release.yml): download and add ManifestReversion.java script to the release workflow
chore(create-release.yml): make ManifestReversion.java script executable in the release workflow

* fix(re-version.sh): fix extraction of tar.gz file by specifying the target directory to avoid creating unnecessary subdirectories
fix(re-version.sh): update version numbers in .txt files to match the new version number

* fix(re-version.sh): fix path issue in re-version.sh script

The path issue in the re-version.sh script was causing the script to fail when trying to copy the liquibase-core.jar and liquibase-commercial.jar files. The fix involves changing the directory to the parent directory before copying the files to the correct location.

* chore(re-version.sh): add debug statements to troubleshoot issues with re-versioning process

* fix(create-release.yml): update script URLs to use the correct branch name

The script URLs in the create-release.yml workflow file were pointing to the wrong branch. This commit updates the URLs to use the correct branch name, which is now stored in the `scripts_branch` variable. This ensures that the correct scripts are downloaded and used during the workflow execution.

* refactor(re-version.sh): remove unnecessary debug statements and change directory context to improve readability and maintainability

* chore(re-version.sh): add debug statements for troubleshooting
fix(re-version.sh): update version number in text files to match the specified version

* fix(re-version.sh): add --strip-components=1 flag to tar command to remove leading directory when extracting tar.gz file

* fix(re-version.sh): remove debug statements and unnecessary file operations to improve script efficiency and readability
fix(create-release.yml): add step to checkout the correct branch or commit for liquibase repository to ensure correct versioning

* chore(create-release.yml): remove unnecessary step to checkout liquibase and liquibase-pro repositories

The step to checkout the liquibase and liquibase-pro repositories was removed as it is no longer needed for the workflow.

---------

Co-authored-by: Sayali Mohadikar <76010603+sayaliM0412@users.noreply.github.com>
Co-authored-by: Sayali M <sayali@Sayalis-MacBook-Pro>