fire update listeners after updating DBCL table (DAT-16772) (#5543)
In an effort to prevent exceptions from listeners causing an out-of-sync situation with the database changelog table, the update listeners are now fired after updating the DBCL table. The concern here is that some listener might throw an exception which causes the write to the DBCL table to not occur, yet the actual changeset contents have already been executed and persisted.

The database object quoting strategy is now being reset before the listeners are fired. If a listener depended on the object quoting strategy being set to some other value, this could break the listener.