add default constructor + setters to `DatabaseChangeLogLock`
this is needed to be able to de-serialise it using [jackson] which
relies on the default constructor & setters.
in the upcoming [liquibase-opensearch] implementation [opensearch-java]
is used, which underneath is using jackson for the (de-)serialisation
of classes.

specifically, this is used in the implementation of
`LockService#listLocks` which needs to load `DatabaseChangeLogLock`s
from the DB.

also do some minor cleanups by removing manually implemented getters &
setters where lombok annotations can do the job instead (no functional
impact).
this could be cleaned up even further if `Date` were replaced with
`LocalTime` (which is immutable), but that'd be a bigger (breaking)
change.

note that this is the same pattern as was previously done in #5336
(commit b4726a0) for `RanChangeSet` & other classes.

[jackson]: https://github.com/FasterXML/jackson
[opensearch-java]: https://github.com/opensearch-project/opensearch-java/
[liquibase-opensearch]: https://github.com/liquibase/liquibase-opensearch/

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>