`CheckSum`: add getter for `storedCheckSum`
this is needed to be able to de-serialise it using [jackson] which
relies on the getters.
in the upcoming [liquibase-opensearch] implementation [opensearch-java]
is used, which underneath is using jackson for the (de-)serialisation
of classes.

this is a follow-up to commit 3823f0bcb (PR #5336) where only the
deserialisation was fixed (by adding a default constructor + setters).

[jackson]: https://github.com/FasterXML/jackson
[opensearch-java]: https://github.com/opensearch-project/opensearch-java/
[liquibase-opensearch]: https://github.com/liquibase/liquibase-opensearch/