[6022] Keep original schema and catalog case
H2 allows users to specify the behavior for casing. This is important when certain databases are mimicked. If the schema and catalog always are rewritten to UPPER_CASE, as it is the default Database.java behavior, situations may occur in which the schema and/or catalog name no longer match the actual database object.

To handle this schema and catalog are now always returned in the original case.